+++ tests/input/no-member/from_unpack.py
@@ -1,0 +1,16 @@
+import pytest
+
+
+def meh():
+    return True, False
+
+
+class TestClass(object):
+    @staticmethod
+    @pytest.fixture(scope='class', autouse=True)
+    def setup_class(request):
+        cls = request.cls
+        cls.defined_in_setup_class, _ = meh()
+
+    def test_foo(self):
+        assert self.defined_in_setup_class